Refactoring and brushing up SimulationServiceImpl and MadratzMatch

+++ server/physics/src/com/madratz/gamelogic/Player.java
@@ -5,24 +5,24 @@ import org.jbox2d.common.Vec2;
 
 public class Player extends Actor {
 
-    private int mId;
+    private long mId;
 
-    protected Player(int id, Behavior behavior) {
+    protected Player(long id, Behavior behavior) {
         super(behavior);
         mId = id;
     }
 
-    public Player(int id, Behavior behavior, Vec2 position, float angle) {
+    public Player(long id, Behavior behavior, Vec2 position, float angle) {
         super(behavior, position, angle);
         mId = id;
     }
 
-    public Player(int id, Vec2 position, float angle) {
+    public Player(long id, Vec2 position, float angle) {
         super(position, angle);
         mId = id;
     }
 
-    public int getId() {
+    public long getId() {
         return mId;
     }
 }
